(#14761) Add boot, reboot to excludes list for redhat provider
On sles, the reboot init script would be triggered during a `puppet resource
service` call, which would ignore the status argument and proceed to reboot the
system. It would also call the boot init script, which could hang the puppet
call indefinitely. This commit adds both the boot and reboot services to the
redhat provider's exclude list. It also updates the redhat provider spec test
to test for those changes.

Signed-off-by: Robert Milasan <rmilasan@suse.com>